@@ -7,8 +7,8 @@
     "serve": "vite preview"
   },
   "dependencies": {
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0"
+    "react": "^18.0.0-alpha-996da67b2-20211018",
+    "react-dom": "^18.0.0-alpha-996da67b2-20211018"
   },
   "devDependencies": {
     "@types/react": "^17.0.0",
